🌿Ensure tests end when Pester run is complete (#130)
* 🐛 End test on pester end (fixes #115)

The test run was only ended when it was cancelled.

* ⬆️ Bump dependencies versions to avoid warnings

npm protests about a security issue.
When packaging the extension, a warning complains about the version of
vsce.

* ✨ Add Command to stop the background  process

When testing a module, with or without classes, it may be necessary to
restart the background process. There is the
`pester.runTestsInNewProcess`  setting for that, but one may not want to
restart the process for each test run. This commit introduces a new
command.

* ✅ Make tests work

Tests were failing for two reasons:

- when `pwsh` is not installed (Vanilla PowerShell 5 on Windows)
- when testing the `PowerShell.exec()` method. AFAIK, this occurs when the execution
  of the PowerShell is too fast

For the first one, I have introduced a constant taking the platform into
account.

For the second one, when check the that the stream is still alive before
waiting for it.