Multi-Root Workspace Support (#173)
This enables multiple root workspaces to exist. There is more to do but this is a good first start. A separate test controller and pwsh instance are started for each workspace so they operate independently and don't clobber each other. PowerShell extension debugging currently all runs through the main terminal window, the goal however is to attach to a process in the future.